refactor(frontend): split the booking chrome so Fast Refresh can keep state
Addresses the ten React Doctor findings on this PR, all in bookingChrome.tsx.

- `only-export-components` x6: the module exported components, class strings
  and pure formatters together, so Fast Refresh had to full-reload rather than
  preserve state. Split three ways - bookingChrome.tsx keeps the components,
  bookingStyles.ts the class recipes, bookingFormat.ts the helpers.
- `js-hoist-intl` x2: both `new Intl.DateTimeFormat()` calls were inside the
  functions, so they were rebuilt on every call. `formatLongDay` backs the day
  hint, which re-renders on every keystroke in the form below it, since they
  share the same component's state. Hoisted to module scope.
- `js-combine-iterations` x1: `groupByDayPart` mapped then filtered, two passes.
  Now one pass that only builds the groups that have slots. Each day part also
  carries its own `from` hour rather than looking back at the previous element,
  so the result no longer depends on the array's order.

React Doctor reports nothing on these files now, and the 47 booking tests pass
unchanged - the split is import paths only, no behaviour.
